add portfolio after buying or selling
