+++ lol_stocks_discord_bot/src/helpers/mod.rs
@@ -1,2 +1,3 @@
 pub mod user_graph_data;
 pub mod plus_sign;
+pub mod portfolio_view;
